python_server/pyproject.toml
[tool.poetry]
name = "python-server"
version = "0.1.0"
description = "Python server for the gaming copilot."
authors = ["David Klueners <davidklueners@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "^0.115.6"
pydantic = "^2.10.5"
pytz = "^2024.2"
uvicorn = "^0.34.0"
requests = "^2.32.3"
markdown2 = "^2.5.2"
python-dotenv = "^1.0.1"
pydantic-ai = "^0.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
